requirements.txt
Flask>=2.2.3
uWSGI>=2.0.21
Flask-Session>=0.4.0
msal>=1.21.0
python-dateutil>=2.8.1
PyYAML>=5.3.1
requests>=2.24.0
sqlalchemy>=2.0.5.post1
python-dateutil>=2.8.1
weasyprint>=58.1
bottle